# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LocalDramaStudio",
    "description": "LOCAL_ONLY local AI short-drama production workbench API",
    "version": "0.1.0-g1"
  },
  "paths": {
...
    "/api/v1/episodes/{episode_id}/timeline-revisions": {
      "post": {
        "tags": [
          "timeline",
          "delivery"
        ],
        "summary": "Create Timeline Revision",
        "operationId": "createTimelineRevision",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Episode Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimelineRevisionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Createtimelinerevision"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/timeline-revisions/{timeline_revision_id}": {
      "get": {
        "tags": [
          "timeline",
          "delivery"
        ],
        "summary": "Get Timeline Revision",
        "operationId": "getTimelineRevision",
        "parameters": [
          {
            "name": "timeline_revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Timeline Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Gettimelinerevision"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/episodes/{episode_id}/subtitle-revisions": {
      "post": {
        "tags": [
          "timeline",
          "delivery"
        ],
        "summary": "Create Subtitle Revision",
        "operationId": "createSubtitleRevision",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Episode Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubtitleRevisionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Createsubtitlerevision"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subtitle-revisions/{subtitle_revision_id}": {
      "get": {
        "tags": [
          "timeline",
          "delivery"
        ],
        "summary": "Get Subtitle Revision",
        "operationId": "getSubtitleRevision",
        "parameters": [
          {
            "name": "subtitle_revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subtitle Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Getsubtitlerevision"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/episodes/{episode_id}/audio-bindings": {
      "post": {
        "tags": [
          "timeline",
          "delivery"
        ],
        "summary": "Bind Episode Audio",
        "operationId": "bindEpisodeAudio",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Episode Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioBindingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Bindepisodeaudio"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "timeline",
          "delivery"
        ],
        "summary": "List Episode Audio",
        "operationId": "listEpisodeAudioBindings",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Episode Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Listepisodeaudiobindings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media-versions/{media_version_id}:create-frame-anchor": {
      "post": {
        "tags": [
          "timeline",
          "delivery"
        ],
        "summary": "Create Frame Anchor",
        "operationId": "createFrameAnchor",
        "parameters": [
          {
            "name": "media_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Version Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FrameAnchorRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Createframeanchor"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/frame-anchors/{anchor_id}": {
      "get": {
        "tags": [
          "timeline",
          "delivery"
        ],
        "summary": "Get Frame Anchor",
        "operationId": "getFrameAnchor",
        "parameters": [
          {
            "name": "anchor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Anchor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Getframeanchor"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shot-transitions": {
      "post": {
        "tags": [
          "timeline",
          "delivery"
        ],
        "summary": "Create Shot Transition",
        "operationId": "createShotTransitionConstraint",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionConstraintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Createshottransitionconstraint"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shot-transitions/{constraint_id}:validate": {
      "post": {
        "tags": [
          "timeline",
          "delivery"
        ],
        "summary": "Validate Shot Transition",
        "operationId": "validateShotTransitionConstraint",
        "parameters": [
          {
            "name": "constraint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Constraint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validateshottransitionconstraint"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/post-process-recipes": {
      "post": {
        "tags": [
          "timeline",
          "delivery"
        ],
        "summary": "Create Post Process Recipe",
        "operationId": "createPostProcessRecipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostProcessRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Createpostprocessrecipe"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/post-process-recipes/{recipe_id}": {
      "get": {
        "tags": [
          "timeline",
          "delivery"
        ],
        "summary": "Get Post Process Recipe",
        "operationId": "getPostProcessRecipe",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recipe Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Getpostprocessrecipe"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/enhancement-runs": {
      "post": {
        "tags": [
          "timeline",
          "delivery"
        ],
        "summary": "Run Enhancement",
        "operationId": "runEnhancement",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnhancementRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Runenhancement"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/timeline-revisions/{timeline_revision_id}:render": {
      "post": {
        "tags": [
          "timeline",
          "delivery"
        ],
        "summary": "Render Episode",
        "operationId": "renderEpisode",
        "parameters": [
          {
            "name": "timeline_revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Timeline Revision Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RenderEpisodeRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Renderepisode"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioBindingRequest": {
        "properties": {
          "media_version_id": {
            "type": "string",
            "minLength": 1,
            "title": "Media Version Id"
          },
          "track_type": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Track Type",
            "default": "AUDIO"
          },
          "start_us": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start Us"
          },
          "end_us": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "End Us"
          },
          "gain_db": {
            "type": "number",
            "title": "Gain Db",
            "default": 0.0
          },
          "source_license_status": {
            "type": "string",
            "maxLength": 24,
            "minLength": 1,
            "title": "Source License Status",
            "default": "VERIFIED_LOCAL"
          }
        },
        "type": "object",
        "required": [
          "media_version_id",
          "start_us",
          "end_us"
        ],
        "title": "AudioBindingRequest"
      },
...
      "EnhancementRunRequest": {
        "properties": {
          "input_media_version_id": {
            "type": "string",
            "minLength": 1,
            "title": "Input Media Version Id"
          },
          "recipe_id": {
            "type": "string",
            "minLength": 1,
            "title": "Recipe Id"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          }
        },
        "type": "object",
        "required": [
          "input_media_version_id",
          "recipe_id"
        ],
        "title": "EnhancementRunRequest"
      },
...
      "FrameAnchorRequest": {
        "properties": {
          "source_time_us": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Time Us"
          },
          "source_frame_index": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Frame Index"
          },
          "position_mode": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(FIRST_FRAME|LAST_FRAME)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Mode"
          },
          "role_hint": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Role Hint",
            "default": "LAST_FRAME"
          }
        },
        "type": "object",
        "title": "FrameAnchorRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PostProcessRecipeRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Code"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "steps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "minItems": 1,
            "title": "Steps"
          },
          "capability_contract": {
            "additionalProperties": true,
            "type": "object",
            "title": "Capability Contract"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "steps"
        ],
        "title": "PostProcessRecipeRequest"
      },
...
      "RenderEpisodeRequest": {
        "properties": {
          "timeline_revision_id": {
            "type": "string",
            "minLength": 1,
            "title": "Timeline Revision Id"
          }
        },
        "type": "object",
        "required": [
          "timeline_revision_id"
        ],
        "title": "RenderEpisodeRequest"
      },
...
      "SubtitleRevisionRequest": {
        "properties": {
          "cues": {
            "items": {
              "$ref": "#/components/schemas/SubtitleCueRequest"
            },
            "type": "array",
            "minItems": 1,
            "title": "Cues"
          },
          "format": {
            "type": "string",
            "maxLength": 16,
            "minLength": 3,
            "title": "Format",
            "default": "SRT"
          },
          "input_snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Snapshot"
          }
        },
        "type": "object",
        "required": [
          "cues"
        ],
        "title": "SubtitleRevisionRequest"
      },
      "TimelineItemRequest": {
        "properties": {
          "track_type": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Track Type",
            "default": "VIDEO"
          },
          "media_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Version Id"
          },
          "start_us": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start Us"
          },
          "end_us": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "End Us"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          }
        },
        "type": "object",
        "required": [
          "start_us",
          "end_us"
        ],
        "title": "TimelineItemRequest"
      },
      "TimelineRevisionRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TimelineItemRequest"
            },
            "type": "array",
            "minItems": 1,
            "title": "Items"
          },
          "input_snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Snapshot"
          },
          "status": {
            "type": "string",
            "maxLength": 24,
            "minLength": 1,
            "title": "Status",
            "default": "DRAFT"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "TimelineRevisionRequest"
      },
      "TransitionConstraintRequest": {
        "properties": {
          "from_shot_id": {
            "type": "string",
            "minLength": 1,
            "title": "From Shot Id"
          },
          "to_shot_id": {
            "type": "string",
            "minLength": 1,
            "title": "To Shot Id"
          },
          "constraint_type": {
            "type": "string",
            "maxLength": 48,
            "minLength": 1,
            "title": "Constraint Type"
          },
          "from_anchor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Anchor Id"
          },
          "to_anchor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Anchor Id"
          },
          "enforcement": {
            "type": "string",
            "maxLength": 16,
            "minLength": 1,
            "title": "Enforcement",
            "default": "HARD"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "from_shot_id",
          "to_shot_id",
          "constraint_type"
        ],
        "title": "TransitionConstraintRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}